+++ components/new-relic/src/actions/events/index.ts
@@ -1,0 +1,4 @@
+import { sendEventData } from "./sendEventData";
+export default {
+  sendEventData,
+};