+++ components/new-relic/src/actions/index.ts
@@ -1,0 +1,10 @@
+import logsActions from "./logs";
+import eventsActions from "./events";
+import metricsActions from "./metrics";
+import miscActions from "./misc";
+export default {
+  ...logsActions,
+  ...eventsActions,
+  ...metricsActions,
+  ...miscActions,
+};