+++ components/new-relic/src/actions/logs/index.ts
@@ -1,0 +1,6 @@
+import { sendLogs } from "./sendLogs";
+import { sendDetailedLogs } from "./sendDetailedLogs";
+export default {
+  sendLogs,
+  sendDetailedLogs,
+};