+++ components/new-relic/src/actions/metrics/index.ts
@@ -1,0 +1,4 @@
+import { sendMetrics } from "./sendMetrics";
+export default {
+  sendMetrics,
+};