+++ components/new-relic/src/connections/index.ts
@@ -1,0 +1,3 @@
+import { apiKeySecret } from "./apiKey";
+export { apiKeySecret };
+export default [apiKeySecret];