+++ components/new-relic/src/index.ts
@@ -1,11 +1,7 @@
 import { component } from "@prismatic-io/spectral";
-import { sendLogs } from "./actions/sendLogs";
-import { sendMetrics } from "./actions/sendMetrics";
-import { sendEventData } from "./actions/sendEvents";
-import { sendDetailedLogs } from "./actions/sendDetailedLogs";
-import connections from "./connections";
-import rawRequest from "./actions/rawRequest";
 import { handleErrors } from "@prismatic-io/spectral/dist/clients/http";
+import actions from "./actions";
+import connections from "./connections";
 export default component({
   key: "new-relic",
   public: true,
@@ -17,13 +13,7 @@ export default component({
       "Send metrics, logs, and events to New Relic observability platform.",
     iconPath: "icon.png",
   },
-  actions: {
-    sendLogs,
-    sendMetrics,
-    sendEventData,
-    sendDetailedLogs,
-    rawRequest,
-  },
+  actions,
   connections,
   hooks: {
     error: handleErrors,