+++ components/new-relic/src/inputs/index.ts
@@ -1,0 +1,5 @@
+export * from "./common";
+export * from "./logs";
+export * from "./metrics";
+export * from "./events";
+export * from "./misc";